move EarlyStop to utils;add verbose control;simplfier the return for func TrainerModel.__str__()
