+++ QCircLearning/trainerModel.py
@@ -7,7 +7,7 @@ import numpy as np
 from .back_minimizer import BackMinimizer
 from scipy.optimize import minimize, OptimizeResult
 from typing import List, Callable
-from .utils import data_augment
+from .utils import data_augment, EarlyStopping
 
 class TrainerModel(nn.Module):
     def __init__(self, layers: List[nn.Module] = None, name: str = None):
@@ -26,8 +26,7 @@ class TrainerModel(nn.Module):
             return pred
 
     def __str__(self):
-        model_structure = self.model.__str__()
-        return f"TrainerModel(name={self.name}):\n{model_structure}"
+        return f"TrainerModel(name={self.name}):\n{self.model}"
 
     def __repr__(self):
         return self.__str__()
@@ -65,28 +64,6 @@ def model_train(model, data_loader, optimizer, device):
     total_loss /= len(data_loader.dataset)
     return total_loss
 
-class EarlyStopping:
-    def __init__(self, patience=5, min_delta=0.0, verbose=False):
-        self.patience = patience
-        self.min_delta = min_delta
-        self.verbose = verbose
-        self.best_loss = None
-        self.counter = 0
-        self.early_stop = False
-
-    def __call__(self, current_loss):
-        if self.best_loss is None:
-            self.best_loss = current_loss
-        elif current_loss < self.best_loss - self.min_delta:
-            self.best_loss = current_loss
-            self.counter = 0
-        else:
-            self.counter += 1
-            if self.verbose:
-                print(f"EarlyStopping counter: {self.counter}/{self.patience}")
-            if self.counter >= self.patience:
-                self.early_stop = True
-
 
 def NN_opt(func, x0, callback=None, **kwargs):
     para_size = len(x0)
@@ -109,28 +86,25 @@ def NN_opt(func, x0, callback=None, **kwargs):
     )
     patience = kwargs.get("patience", 5)
     min_delta = kwargs.get("min_delta", 0.0)
-    print("Patience: ", patience)
-    print("Min Delta: ", min_delta)
 
     early_stopping = EarlyStopping(patience=patience, min_delta=min_delta,verbose=verbose)
-    # print(f"Early stopping: {early_stopping}")
-    print(f"patience: {early_stopping.patience}, min_delta: {early_stopping.min_delta}, verbose: {early_stopping.verbose}")
+    
     sample_x = init_data
     sample_y = [func(para) for para in sample_x]
     optimal = [sample_x[np.argmin(sample_y)], np.min(sample_y)]
-    print(f"Training with the neural networks")
+    if verbose:
+        print(f"Training with the neural networks")
     sys.stdout.flush()
 
     for model in nn_models:
-        print(model)
-        sys.stdout.flush()
+        if verbose:
+            print(model)
+            sys.stdout.flush()
 
         optimizer = optim.Adam(model.parameters(), lr=kwargs.get("lr",1e-4))
 
         for iteration in range(max_iter):
             res.nit += 1
-            print(f"Iteration {iteration + 1}/{max_iter}")
-            print(f"batch_size: {batch_size}")
             data_loader = DataLoader(
                 list(zip(sample_x, sample_y)), batch_size=batch_size, shuffle=True
             )
@@ -139,7 +113,11 @@ def NN_opt(func, x0, callback=None, **kwargs):
                 total_loss = model_train(model, data_loader, optimizer, device)
                 early_stopping(total_loss)
                 if early_stopping.early_stop:
-                    print(f"Early stopping triggered at iteration {iteration + 1}, epoch {epoch + 1}")
+                    if verbose:
+                        print(
+                            f"Early stopping at epoch {epoch + 1}/{classical_epochs}, Average Loss: {total_loss:.1e}"
+                        )
+                        sys.stdout.flush()
                     break
                 if verbose:
                     print(
@@ -161,9 +139,9 @@ def NN_opt(func, x0, callback=None, **kwargs):
 
             augment_points = kwargs.get("augment_points", 2)
             predictions = data_augment(prediction0, augment_points, shift=np.pi * 2)
-            for pred in predictions:
-                if not np.isfinite(func(pred)):  # Check if `func` can handle the augmented data
-                    print(f"Invalid prediction: {pred}")
+            # for pred in predictions:
+            #     if not np.isfinite(func(pred)):  # Check if `func` can handle the augmented data
+            #         print(f"Invalid prediction: {pred}")
 
             sample_x += predictions
             sample_y += [y0] * len(predictions)
@@ -182,12 +160,12 @@ def random_search(func, x0, callback=None, **kwargs):
         "init_data", [np.random.uniform(-10, 10, para_size) for _ in range(60)]
     )
     max_iter = kwargs.get("max_iter", 20)
-
+    verbose = kwargs.get("verbose", 0)
     sample_x = init_data
     sample_y = [func(para) for para in sample_x]
     optimal = [sample_x[np.argmin(sample_y)], np.min(sample_y)]
-
-    print("Training with random search")
+    if verbose:
+        print("Training with random search")    
     sys.stdout.flush()
 
     for _ in range(max_iter):
